cast usize to u32

+++ crates/oxc_ast/src/ast_builder_impl.rs
@@ -43,8 +43,8 @@ impl<'a> AstBuilder<'a> {
     }
 
     #[inline]
-    pub fn vec_with_capacity<T>(self, capacity: usize) -> Vec<'a, T> {
-        Vec::with_capacity_in(capacity as u32, self.allocator)
+    pub fn vec_with_capacity<T>(self, capacity: u32) -> Vec<'a, T> {
+        Vec::with_capacity_in(capacity, self.allocator)
     }
 
     #[inline]
